try to catch exceptions in threads too, not just the main thread. fix default settings to correctly use ` instead of ~ for toggleconsole. fix a couple of possible signed-where-unsigned-expected with quake2. fix possible q3 crash.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4649 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_parse.c
@@ -4504,7 +4504,7 @@ void CLQ2_ParseMuzzleFlash (void)
 	float		volume;
 	char		soundname[64];
 
-	i = MSG_ReadShort ();
+	i = (unsigned short)(short)MSG_ReadShort ();
 	if (i < 1 || i >= Q2MAX_EDICTS)
 		Host_Error ("CL_ParseMuzzleFlash: bad entity");
 
@@ -4692,7 +4692,7 @@ void CLQ2_ParseMuzzleFlash2 (void)
 	int			ent;
 	int			flash_number;
 
-	ent = MSG_ReadShort ();
+	ent = (unsigned short)(short)MSG_ReadShort ();
 	if (ent < 1 || ent >= Q2MAX_EDICTS)
 		Host_EndGame ("CL_ParseMuzzleFlash2: bad entity");
